Made own compilation unit for rcd_demosaic()
